refactor(db): move thread store ownership
